chore: Exception handling update

+++ src/main/java/br/com/chaves/course/services/UserService.java
@@ -4,11 +4,11 @@ import br.com.chaves.course.entities.User;
 import br.com.chaves.course.repositories.UserRepository;
 import br.com.chaves.course.services.exceptions.DatabaseException;
 import br.com.chaves.course.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.ResourceAccessException;
 
 import java.util.List;
 import java.util.Optional;
@@ -31,20 +31,36 @@ public class UserService {
         return repository.save(obj);
     }
 
+//    public void delete(Long id) {
+//        try {
+//            repository.deleteById(id);
+//        } catch (EmptyResultDataAccessException e) {
+//            throw new ResourceNotFoundException(id);
+//        } catch (DataIntegrityViolationException e) {
+//            throw new DatabaseException(e.getMessage());
+//        }
+//    }
+//
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
         try {
             repository.deleteById(id);
-        } catch (EmptyResultDataAccessException e){
-            throw new ResourceNotFoundException(id);
-        } catch (DataIntegrityViolationException e){
+        } catch (DataIntegrityViolationException e) {
             throw new DatabaseException(e.getMessage());
         }
     }
 
     public User update(Long id, User obj) {
-        User entity = repository.getReferenceById(id);
-        updateData(entity, obj);
-        return repository.save(entity);
+        try {
+            User entity = repository.getReferenceById(id);
+            updateData(entity, obj);
+            return repository.save(entity);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User obj) {
